requirements.txt
discord.py==2.5.0
environs==14.1.1
